Create endpoint API for adding new question.
Create endpoint api for create new Question.
In POSTMAN, it will be a POST.
with Headers: Accept: application/json
Authorization: Bearer + accesstoken.
Accesstoken is what we have when POST to api/token. +(with body is: username +password).
Then with that token, in Create question, we need the body of the POST is: "title", "body"