Fixing Delete question button.
User case:
As an user, I want when a Question does not have any Answer, it will be deleteable.
But when it has at least an Answer, it is not able to delete Question.
And the button must also dynamic.
Solution: add $emit to Event bus.
When Answer is add/remove it will $emit a signal to Question, Question will count it's Answers and decide the Delete button to appear or not/.